Move events-current-list component to components directory
