+++ src/app/events/components/current/events-current-list/events-current-list.component.spec.ts
