+++ src/app/events/components/current/events-current-list/events-current-list.component.ts
@@ -1,6 +1,6 @@
 import { ChangeDetectionStrategy, Component } from "@angular/core";
 import { TranslatePipe } from "@ngx-translate/core";
-import { EventsListComponent } from "../../components/common/events-list/events-list.component";
+import { EventsListComponent } from "../../common/events-list/events-list.component";
 
 @Component({
   selector: "bkd-events-current-list",
